controller: set CronJob.spec.SuccessfulJobsHistoryLimit to 0
CronJob leaves the last three jobs by default. When the controller's pod
is restarted, CronJobs are deleted due to the ownerReferences. The three
remaining jobs increase the number of resources to be deleted at the
same time, and this may cause a high load on the api server.

This patch changes CronJob so that successful jobs are not left behind.

Co-authored-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>
Signed-off-by: Daichi Mukai <daichi-mukai@cybozu.co.jp>