observer: delete Job instead of Pod after create threshold
Pie has the ability to delete Pods that have not started within a
certain period of time. If a Pod does not start, there may be a problem,
and we do not want to have Pods being created over and over again.

In this commit, we change so that if a Pod does not start within a
certain period of time, the Jobs in the ownerReferences will be deleted
instead of the Pod.

Co-authored-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>
Co-authored-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>
Signed-off-by: Daichi Mukai <daichi-mukai@cybozu.co.jp>