Remove ownerReference
To ensure that cronjobs are deleted when the pie is deleted,
ownerReferences to the pie pod were set in the cronjobs. However, when a
pie update occurred, the old pod was deleted and the cronjobs were
deleted simultaneously. This deletion was causing high loads.

In this PR, the ownerReference was deleted to avoid high loads.

Co-authored-by: Daichi Mukai <daichi-mukai@cybozu.co.jp>
Co-authored-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>
Signed-off-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>