charts, test/e2e: synchronize PieProbe CRD
In pull request #101, we introduced a new custom resource called
PieProbe. However, the charts/ and test/e2e/ directories were not
synchronized with the auto-generated config/ directory.

This patch fixes the above probelm, by copying the latest PieProbe CRD
from config/ to charts/ and updating the manifests used in the e2e
tests.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>