fix provision_observer2 to delete failing Jobs correctly
The new provision observer (provision_observer2.go) introduced in #101 tries to
remove the owned jobs that is failing for a while, but it doesn't work well
because it looks for them by a wrong prefix.

This PR fixes the above problem by using constants.MountProbeNamePrefix and
constants.ProvisionProbeNamePrefix instead of constants.ProbeNamePrefix when
the provision observer looks for the jobs.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>