fix: use at most 15 letters for node name in cronjob name
When the current implementation decides a cronjob name, it uses at most
11 letters for a node name. This behavior is problematic when a node
name is an IP address, because it can have at most 15 (= 3*4+3)
letters. This commit fixes this problem by including at most 15
letters of the node name in the cronjob name. In compensation, this
commit shrinks the length of the storage class name to 12.

A mount cronjob name (at most 40 letters) consists of:

- "provision-" (10 letters)
- PieProbe resource name (at most 10 letters)
- hyphen (1 letter)
- StorageClass name (at most 12 letters)
- hyphen (1 letter)
- hash (6 letters)

A provision cronjob name (at most 52 letters) consists of:

- "mount-" (6 letters)
- PieProbe resource name (at most 10 letters)
- hyphen (1 letter)
- Node name (at most 15 letters)
- hyphen (1 letter)
- StorageClass name (at most 12 letters)
- hyphen (1 letter)
- hash (6 letters)

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>